requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# AI/LLM
openai==1.3.0
anthropic==0.7.0
langchain==0.0.350
tiktoken==0.5.1

# Data processing
pandas==2.1.3
numpy==1.26.2

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23
reportlab==4.0.7

# Web framework
requests==2.31.0
httpx==0.25.1

# Database (optional)
sqlalchemy==2.0.23
sqlite3

# Utilities
python-dateutil==2.8.2
pytz==2023.3

